added

+++ msip-bhu-business-impl/msip_bhu_rpc_api/src/main/java/com/bhu/vas/api/vto/WifiDeviceMaxBusyVTO.java
@@ -2,7 +2,7 @@ package com.bhu.vas.api.vto;
 
 public class WifiDeviceMaxBusyVTO {
 	private String wid;//wifi id
-	private int hdc;//handset device count 移动设备接入的总数量
+	private long hdc;//handset device count 移动设备接入的总数量
 	
 	public String getWid() {
 		return wid;
@@ -10,10 +10,10 @@ public class WifiDeviceMaxBusyVTO {
 	public void setWid(String wid) {
 		this.wid = wid;
 	}
-	public int getHdc() {
+	public long getHdc() {
 		return hdc;
 	}
-	public void setHdc(int hdc) {
+	public void setHdc(long hdc) {
 		this.hdc = hdc;
 	}
 }
